--- src/main/java/frc/robot/commands/ClimbingMotorsSetSpeed.java
@@ -1,40 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.ClimbingSubsystem;
-
-public class ClimbingMotorsSetSpeed extends Command {
-  private final ClimbingSubsystem climbingSubsystem = ClimbingSubsystem.getInstance();
-  double speed;
-  /** Creates a new ClimbingMotorsSetSpeed. */
-  public ClimbingMotorsSetSpeed(double speed) {
-    this.addRequirements(climbingSubsystem);
-    this.speed = speed;
-
-    // Use addRequirements() here to declare subsystem dependencies.
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    this.climbingSubsystem.setSpeedClimbing(speed);
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {}
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {}
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
